requirements.txt
streamlit
google-generativeai
python-dotenv
streamlit-lottie
streamlit-card
requests